modify timed_tasks

+++ internal/app/models/procedure.go
@@ -14,7 +14,7 @@ type Procedure struct {
 	State              uint       `gorm:"not null" json:"state"`
 	PHash              string     `gorm:"type:varchar(256)" json:"p_hash"`
 	CheckCode          string     `gorm:"type:varchar(256)" json:"check_code"`
-	SerialNumber       uint       `gorm:"not null" json:"serial_number"`
+	SerialNumber       int64      `gorm:"not null" json:"serial_number"`
 	Operator           string     `json:"operator"`
 	StartTimestamp     time.Time  `gorm:"not null" json:"start_timestamp"`
 	CompletedTimestamp *time.Time `json:"completed_timestamp"`
